apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "djangorestframework>=3.15.0",
    "drf-spectacular>=0.27.0",
    "django-filter>=24.0",
    "gunicorn>=23.0.0",
    "ruff>=0.12.0",
    "dj-database-url>=2.1.0",
    "psycopg2-binary>=2.9.9",
    "mypy>=1.8.0",
    "django-stubs>=4.2.0",
    "djangorestframework-stubs>=3.14.0",
    "jinja2>=3.1.0",
    "django-vite>=3.1.0",
    "djangorestframework-simplejwt>=5.3.0",
]

[build-system]
requires = ["uv_build>=0.8.2,<0.10.0"]
build-backend = "uv_build"


[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "settings"
